🧹 [Tournament] Extract UI state to custom hook
